Add PartyHeader to Manifesto page
