+++ app/assets/stylesheets/index.scss
@@ -6,5 +6,5 @@
 @import './pages/home/home';
 @import './pages/party-detail/party-detail';
 @import './pages/party-candidates/party-candidates';
-@import './pages/party-program/party-program';
+@import './pages/party-manifesto/party-manifesto';
 @import './pages/about-us/about-us';
