+++ app/javascript/components/party/index.js
@@ -59,7 +59,7 @@ export default class Party extends PureComponent {
             <Layout>
                 <LayoutHeader />
                 <Layout.Content className="party-section">
-                    <PartyHeader party={party} subtitle={party.acronym} />
+                    <PartyHeader party={party} subtitle={party.acronym} showManifestoLink={true} />
                     <PartyIntro spokesperson={spokesperson} title="Descrição do Partido">
                         <Paragraph className="party-desc">{party.description}</Paragraph>
                         {party.description_source.split('\n').map((item, i) => <Paragraph key={i}>Fonte: <a href={item} target="_blank" rel="noopener">Wikipedia</a></Paragraph>)}
